Trim down specs to only test simple conditions
We don't really need to test the happy path or the specific UTF-8
versions of when a payload is too large because the happy path is tested
in other specs and the UTF-8 issue is really just testing the `bytesize`
method. Also we don't really need to expose the `payload_to_large?`
method so I made that private and removed the tests.