Memoize the Feedback stream
The underlying stream is a socket, meaning we can only read forward and
that reading from it mutates. The net result is if we `#count` the
number of Feedback items, we read them all off the socket and they are
no longer there when we go to enumerate over them later. (See #80 for
more info) This is an experiment to memoize the results of the reading
the socket.

Ultimately, I believe this is a Bad Idea™ because it means the
`Feedback` object can only be read from once. This is a change from the
current behavior where you could re-use the same `Feedback` object (and
hence, connection to APNS) occasionally checking it for new feedback.