Encouraging developer to lock and encrypt the private key.
The current README provides instructions on how to export the certificate
and private key from Keychain Access and prepare it to be used by Grocer.
Unfortunately, the openssl command provided to convert the file from the
`.p12` format to the `.pem` format generates a file where the private key
is left unprotected.

This PR adds the  option to the sample command to encourage developers
to encrypt and lock the private key.